% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure}
\alias{exposure}
\title{The main function of the package.}
\usage{
exposure(d)
}
\arguments{
\item{d}{tibble.  See Details below. Required}
}
\description{
Calculating exposure over multiple referrals
}
