% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-parameter-check.R
\name{exposure_parameter_check}
\alias{exposure_parameter_check}
\title{Check the parameters to exposure}
\usage{
exposure_parameter_check(d, censored_date)
}
\arguments{
\item{d}{tibble.  See Details below. Required}

\item{censored_date}{Date.  The last observed date; this value calculates the censored duration. Required}
}
\description{
Make sure the parameters are what the \code{exposure()} function expects.
}
