% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferralExposure-package.R
\docType{package}
\name{ReferralExposure-package}
\alias{ReferralExposure-package}
\alias{ReferralExposure}
\title{Calculating exposure over multiple referrals}
\description{
Much of this package has been developed to support the needs of the following projects.  We appreciate the support.
\enumerate{

  \item \emph{Title IV-E Waiver Project}, [HRSA/MCHB](http://mchb.hrsa.gov/)-sponsored collaboration with \href{http://www.okdhs.org/}{OKDHS}; David Bard, PI, OUHSC; 2014-2017.

  \item \emph{Shelter Project}, ; David Bard, PI; 2016-2017.

  \item Additional Insitutional Support from OUHSC \href{https://www.oumedicine.com/department-of-pediatrics}{Dept of Pediatrics}; 2013-2017.

}
}
\note{
The most recent development version is available through \href{https://github.com/OuhscBbmc/referral-exposure}{GitHub} by
running
\code{devtools::install_github} \code{('OuhscBbmc/referral-exposure')}
(make sure \href{https://cran.r-project.org/package=devtools}{devtools} is already installed).
If you're having trouble with the package, please install the development version.  If this doesn't solve
your problem, please create a \href{https://github.com/OuhscBbmc/referral-exposure/issues}{new issue}, or email Will.
}
\examples{
\dontrun{
# Install/update ReferralExposure with the development version from GitHub
#install.packages('devtools') #Uncomment if `devtools` isn't installed already.
devtools::install_github('OuhscBbmc/referral-exposure')
}
}
